Moved SwordError and UriRegistry out of the config package, as they have nothing to do with capturing the config.yml
