+++ src/main/java/nl/knaw/dans/sword2/core/UriRegistry.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.sword2.config;
+package nl.knaw.dans.sword2.core;
 
 import javax.ws.rs.core.MediaType;
 import javax.xml.namespace.QName;
