+++ src/main/java/nl/knaw/dans/sword2/resources/ContainerResourceImpl.java
@@ -15,7 +15,7 @@
  */
 package nl.knaw.dans.sword2.resources;
 
-import nl.knaw.dans.sword2.config.SwordError;
+import nl.knaw.dans.sword2.core.SwordError;
 import nl.knaw.dans.sword2.core.auth.Depositor;
 import nl.knaw.dans.sword2.core.exceptions.CollectionNotFoundException;
 import nl.knaw.dans.sword2.core.exceptions.DepositNotFoundException;
